Few fixes for ids

+++ src/A3DB/Workload.hpp
@@ -19,6 +19,6 @@ public:
 	float id;
 	std::string WorkloadCategory;
 	std::string WorkloadData;
-	Workload(int _id, const char *category, std::string request) : id(_id), WorkloadCategory(category), WorkloadData(request) {}
+	Workload(float _id, const char *category, std::string request) : id(_id), WorkloadCategory(category), WorkloadData(request) {}
 };
 #endif
